refactor: separate styled-components and utils in BatchOperations components
Move inline styled-components into sibling styled.ts files for
BatchStateIndicator and BatchItemsCount, and extract formatOperationType and
formatStartDate from BatchOperations into a utils module, so each file holds a
single concern as the file organization conventions require. No behavior change.
