+++ webapp/client/apps/orchestration-cluster-webapp/src/operate/shared/BatchStateIndicator/styled.ts
@@ -1,0 +1,20 @@
+/*
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Camunda License 1.0. You may not use this file
+ * except in compliance with the Camunda License 1.0.
+ */
+
+import styled from 'styled-components';
+
+const Container = styled.div`
+	display: flex;
+
+	svg {
+		align-self: center;
+		margin-inline-end: var(--cds-spacing-03);
+	}
+`;
+
+export {Container};
